# [pins-workflow] Improve GitHub continuous integration performance.