--- a/tests/integration/system/nsf/traffic_helpers/otg_helper.h
+++ b/tests/integration/system/nsf/traffic_helpers/otg_helper.h
@@ -22,8 +22,6 @@
 namespace pins_test {
 
 class OtgHelper : public TrafficHelper {
-  ~OtgHelper() override = default;
-
  public:
   absl::Status StartTraffic(thinkit::GenericTestbed& testbed) override {
     return absl::OkStatus();